+++ rust/mlt-nom/Cargo.toml
@@ -10,9 +10,14 @@ categories.workspace = true
 rust-version.workspace = true
 
 [dependencies]
+borrowme = "0.0.15"
 integer-encoding = "4.0.2"
 nom = "8.0.0"
 num_enum = "0.7.4"
+#mvt-reader = "2.1.0"
+galileo-mvt = "0.2.1"
+mvt = "0.10.3"
+walkdir = "2.5.0"
 
 [lints.rust]
 unsafe_code = "forbid"
@@ -22,3 +27,4 @@ unused_qualifications = "warn"
 cargo = { level = "warn", priority = -1 }
 pedantic = { level = "warn", priority = -1 }
 missing_errors_doc = "allow"
+multiple_crate_versions = "allow"
